liquidationBot - add support for deployment v4.1. WIP
there are still leftovers:
 * liquidationBotApi address is missing
 * docs
 * set default deployment to 4.1
