+++ src/config/common.ts
@@ -9,8 +9,10 @@ import { JsonRpcProvider, Provider } from "@ethersproject/providers";
 import { Wallet } from "@ethersproject/wallet";
 import { IExchangeEvents__factory } from "@generated/factories/IExchangeEvents__factory";
 import { IExchange__factory } from "@generated/factories/IExchange__factory";
+import { TradeRouter__factory } from "@generated/factories/TradeRouter__factory";
 import { IExchange } from "@generated/IExchange";
 import { IExchangeEvents } from "@generated/IExchangeEvents";
+import { TradeRouter } from "@generated/TradeRouter";
 import { Arguments, Argv } from "yargs";
 import { getEnumArg, getNumberArg, getStringArg } from "./args";
 
@@ -24,10 +26,8 @@ export function checkDefined<T>(
   return val;
 }
 
-export enum Network {
-  RINKEBY_ARBITRUM = "rinkeby_arbitrum",
-  MAINNET_ARBITRUM = "mainnet_arbitrum",
-}
+const networksAvailable = ["RINKEBY_ARBITRUM", "MAINNET_ARBITRUM"] as const;
+export type Network = typeof networksAvailable[number];
 
 export type WithNetworkArgs<T = {}> = T & { network: string | undefined };
 export const withNetworkArgv = <T = {}>(
@@ -52,7 +52,7 @@ export const getNetwork = <T = {}>(
   const network = getEnumArg(
     "network",
     "NETWORK",
-    Object.values(Network),
+    Object.values(networksAvailable),
     argv,
     { ignoreCase: true }
   );
@@ -154,7 +154,7 @@ export const getSigner = <T = {}>(
   return { network, signer };
 };
 
-export type ExchangeArgs<T = {}> = WithProviderArgs<
+export type ExchangeArgs<T = {}> = WithSignerArgs<
   T & { exchange: string | undefined }
 >;
 export const exchangeWithProviderArgv = <T = {}>(
@@ -164,7 +164,22 @@ export const exchangeWithProviderArgv = <T = {}>(
     describe:
       "Address of the exchange to interact with.\n" +
       ".env property: <network>_EXCHANGE\n" +
-      "Required",
+      "Required unless deployment-version is 4.1. Ignored othervice",
+    type: "string",
+  });
+};
+
+export type TradeRouterArgs<T = {}> = WithSignerArgs<
+  T & { "trader-router": string | undefined }
+>;
+export const traderRouterWithProviderArgv = <T = {}>(
+  yargs: Argv<T>
+): Argv<TradeRouterArgs<T>> => {
+  return withSignerArgv(yargs).option("trader-router", {
+    describe:
+      "Address of the exchange to interact with.\n" +
+      ".env property: <network>_EXCHANGE\n" +
+      "Required if deployment-version is 4.1. Ignored othervice",
     type: "string",
   });
 };
@@ -175,7 +190,7 @@ export type GetExchangeWithSignerArgv<T> = Arguments<
 export const getExchangeWithSigner = <T = {}>(
   argv: GetExchangeWithSignerArgv<T>
 ): {
-  network: string;
+  network: Network;
   signer: Signer;
   exchangeAddress: string;
   exchange: IExchange;
@@ -191,11 +206,36 @@ export const getExchangeWithSigner = <T = {}>(
   return { network, signer, exchangeAddress, exchange, exchangeEvents };
 };
 
+export type GetTradeRouterWithSignerArgv<T> = Arguments<
+  WithSignerArgs<TradeRouterArgs<T>>
+>;
+export const getTradeRouterWithSigner = <T = {}>(
+  argv: GetTradeRouterWithSignerArgv<T>
+): {
+  tradeRouter: TradeRouter;
+  tradeRouterEvents: IExchangeEvents;
+  tradeRouterAddress: string;
+  signer: Signer;
+} => {
+  const { network, signer } = getSigner(argv);
+  const tradeRouterAddress = getStringArg(
+    "trader-router",
+    `${network}_TRADE_ROUTER`,
+    argv
+  );
+  const tradeRouter = TradeRouter__factory.connect(tradeRouterAddress, signer);
+  const tradeRouterEvents = IExchangeEvents__factory.connect(
+    tradeRouterAddress,
+    signer
+  );
+  return { signer, tradeRouter, tradeRouterEvents, tradeRouterAddress };
+};
+
 export type GetExchangeWithProviderArgv<T> = Arguments<ExchangeArgs<T>>;
 export const getExchangeWithProvider = <T = {}>(
   argv: GetExchangeWithProviderArgv<T>
 ): {
-  network: string;
+  network: Network;
   provider: Provider;
   exchangeAddress: string;
   exchange: IExchange;
